move process .py tests to composites/toys.py. increase pytest coverage of processes.py
