+++ vivarium/composites/toys.py
@@ -1,9 +1,11 @@
 import math
-from typing import Optional, Dict, Any
+from typing import Optional, Dict, Any, Union
 
 import numpy as np
 
-from vivarium.core.process import Process, Deriver, Composer
+from vivarium.core.process import (
+    Process, Deriver, Composer, Composite, AggregateComposer)
+from vivarium.core.types import State, Schema, Update, Topology
 
 quark_colors = ['green', 'red', 'blue']
 quark_spins = ['up', 'down']
@@ -466,7 +468,7 @@ class PoQo(Composer):
         }
 
 
-def test_composite_initial_state() -> None:
+def test_composite_initial_state_complex() -> None:
 
     outer_path = ('universe', 'agent')
     pq = PoQo({})
@@ -484,5 +486,286 @@ def test_composite_initial_state() -> None:
     assert pq_initial == expected_initial
 
 
+class AA(Process):
+    name = 'AA'
+
+    def initial_state(self, config: Optional[dict] = None) -> State:
+        return {'a_port': {'a': 1}}
+
+    def ports_schema(self) -> Schema:
+        return {'a_port': {'a': {'_emit': True}}}
+
+    def next_update(
+            self,
+            timestep: Union[float, int],
+            states: State) -> Update:
+        return {'a_port': {'a': 1}}
+
+
+class BB(Composer):
+    name = 'BB'
+
+    def generate_processes(
+            self, config: Optional[dict]) -> Dict[str, Any]:
+        return {
+            'a1': AA({}),
+            'a2': AA({}),
+            'a3_store': {
+                'a3': AA({})}}
+
+    def generate_topology(self, config: Optional[dict]) -> Topology:
+        return {
+            'a1': {
+                'a_port': ('a1_store',)},
+            'a2': {
+                'a_port': {
+                    'a': ('a1_store', 'b')}},
+            'a3_store': {
+                'a3': {
+                    'a_port': ('a3_1_store',)}}}
+
+
+class ToyProcess(Process):
+    name = 'toy'
+
+    def ports_schema(self) -> Schema:
+        return {
+            'A': {
+                'a': {'_default': 0},
+                'b': {'_default': 0}},
+            'B': {
+                'a': {'_default': 0},
+                'b': {'_default': 0}}}
+
+    def next_update(
+            self, timestep: Union[float, int], states: State) -> Update:
+        return {
+            'A': {
+                'a': 1,
+                'b': states['A']['a']},
+            'B': {
+                'a': states['A']['b'],
+                'b': states['B']['a']}}
+
+
+class ToyComposer(Composer):
+    defaults = {
+        'A':  {'name': 'A'},
+        'B': {'name': 'B'}}
+
+    def generate_processes(
+            self,
+            config: Optional[dict]
+    ) -> Dict[str, ToyProcess]:
+        assert config is not None
+        config = config or self.defaults
+        a = ToyProcess(config['A'])
+        b = ToyProcess(config['B'])
+        return {
+            a.name: a,
+            b.name: b}
+
+    def generate_topology(
+            self,
+            config: Optional[dict] = None
+    ) -> Topology:
+        config = config or self.defaults
+        a_name = config['A']['name']
+        b_name = config['B']['name']
+        return {
+            a_name: {
+                'A': ('aaa',),
+                'B': ('bbb',)},
+            b_name: {
+                'A': ('bbb',),
+                'B': ('ccc',)}}
+
+
+def test_override() -> None:
+    config = {
+        '_schema': {
+            'a3_store': {
+                'a3': {
+                    'a_port': {
+                        'a': {
+                            '_default': 2}}}}}}
+    bb_composer = BB(config)
+    bb_composite = bb_composer.generate()
+    default_state = bb_composite.default_state()
+
+    expected_default_state = {
+        'a3_store': {
+            'a3_1_store': {
+                'a': 2}}}
+
+    assert default_state == expected_default_state
+
+
+def test_composite_initial_state() -> None:
+    """
+    test that initial state in composite merges individual processes'
+    initial states
+    """
+
+    bb_composer = BB({})
+    bb_composite = bb_composer.generate()
+
+    composer_initial_state = bb_composer.initial_state()
+    composite_initial_state = bb_composite.initial_state()
+
+    expected_initial_state = {
+        'a3_store': {
+            'a3_1_store': {
+                'a': 1}},
+        'a1_store': {
+            'a': 1,
+            'b': 1}}
+
+    assert composite_initial_state == composer_initial_state
+    assert composite_initial_state == expected_initial_state
+
+
+def test_composite_parameters() -> None:
+    """
+    test that initial state in composite merges individual processes'
+    initial states
+    """
+
+    bb_composer = BB({})
+    bb_composite = bb_composer.generate()
+    composer_parameters = bb_composer.get_parameters()
+    composite_parameters = bb_composite.get_parameters()
+    expected_parameters = {
+        'a1': {'time_step': 1.0},
+        'a2': {'time_step': 1.0},
+        'a3_store': {
+            'a3': {'time_step': 1.0}}}
+    assert composite_parameters == composer_parameters
+    assert composite_parameters == expected_parameters
+
+
+def test_composite_merge() -> None:
+    composer = ToyComposer()
+    composite = composer.generate()
+
+    expected_initial_topology = {
+        'A': {
+            'A': ('aaa',),
+            'B': ('bbb',),
+        },
+        'B': {
+            'A': ('bbb',),
+            'B': ('ccc',),
+        },
+    }
+    assert composite['topology'] == expected_initial_topology
+
+    for key in ('A', 'B'):
+        assert key in composite['processes']
+        assert isinstance(composite['processes'][key], ToyProcess)
+
+    # merge
+    merge_processes: Dict[str, Process] = {
+        'C': ToyProcess({'name': 'C'})}
+    merge_topology: Topology = {
+        'C': {
+            'A': ('aaa',),
+            'B': ('bbb',)}}
+    composite.merge(
+        processes=merge_processes,
+        topology=merge_topology)
+
+    expected_merged_topology = {
+        'A': {
+            'A': ('aaa',),
+            'B': ('bbb',),
+        },
+        'B': {
+            'A': ('bbb',),
+            'B': ('ccc',),
+        },
+        'C': {
+            'A': ('aaa',),
+            'B': ('bbb',),
+        },
+    }
+    assert composite['topology'] == expected_merged_topology
+
+    for key in ('A', 'B', 'C'):
+        assert key in composite['processes']
+        assert isinstance(composite['processes'][key], ToyProcess)
+
+
+def test_get_composite() -> None:
+    a = ToyProcess({'name': 'a'})
+
+    composite = Composite(dict(
+        processes={'b': ToyProcess()},
+        topology={'b': {'A': ('A',), 'B': ('B',)}}
+    ))
+    composite.merge(
+        processes={a.name: a},
+        topology=a.generate_topology())
+
+    expected_topology = {
+        'a': {
+            'A': ('A',),
+            'B': ('B',)},
+        'b': {
+            'A': ('A',),
+            'B': ('B',)}}
+
+    assert composite['topology'] == expected_topology
+
+
+def test_aggregate_composer() -> None:
+    config1 = {'name': 'one'}
+    aggregate = AggregateComposer(
+        composers=[ToyComposer(config1)])
+    composite1 = aggregate.generate()
+
+    # add composers (list)
+    config2 = {
+        'name': 'two',
+        'A':  {'name': 'AA'},
+        'B': {'name': 'BB'},
+    }
+    aggregate.add_composers(
+        composers=[ToyComposer(config2)],
+        config={})
+    composite2 = aggregate.generate()
+
+    # add composer (single)
+    config3 = {
+        'name': 'three',
+        'A':  {'name': 'AAA'},
+        'B': {'name': 'BBB'},
+    }
+    aggregate.add_composer(
+        composer=ToyComposer(config3),
+        config={})
+    composite3 = aggregate.generate()
+
+    assert all(
+        item in composite2['processes'].keys()
+        for item in composite1['processes'].keys())
+    assert all(
+        item in composite3['processes'].keys()
+        for item in composite2['processes'].keys())
+    assert all(
+        item in composite2['topology'].keys()
+        for item in composite1['topology'].keys())
+    assert all(
+        item in composite3['topology'].keys()
+        for item in composite2['topology'].keys())
+    assert len(composite1['processes']) < len(composite2['processes'])
+    assert len(composite2['processes']) < len(composite3['processes'])
+
 if __name__ == '__main__':
-    test_composite_initial_state()
+    test_composite_initial_state()  # pragma: no cover
+    test_composite_parameters()  # pragma: no cover
+    test_composite_merge()  # pragma: no cover
+    test_get_composite()  # pragma: no cover
+    test_aggregate_composer()  # pragma: no cover
+    test_override()  # pragma: no cover
+    test_composite_initial_state_complex()  # pragma: no cover
